Board, Post Views

+++ everytime/post/views.py
@@ -8,6 +8,8 @@ from django.core.exceptions import ObjectDoesNotExist
 
 from post.models import Post, UserLikePost
 from post.serializers import PostSerializer
+from board.models import Board
+from user.models import UserProfile
 
 # GET /find/board/article/list | list posts
 # POST /save/board/article | write a post
@@ -17,11 +19,12 @@ from post.serializers import PostSerializer
 class PostViewSet(viewsets.GenericViewSet):
     queryset = Post.objects.all()
     serializer_class = PostSerializer
-    #permission_classes = (IsAuthenticated(),)
+    permission_classes = (IsAuthenticated(), )
 
-    # get_serializer_class() ?
-    # UserLikePostSerializer ?
-    #TODO: check user permission
+    def get_permissions(self):
+        if self.action in ('listAllPosts', 'listPosts'):
+            return (AllowAny(), )
+        return self.permission_classes
 
     @action(detail=False, methods=['GET'], url_path='listall')
     def listAllPosts(self, request): # GET /api/post/listall/ | list all posts
@@ -30,17 +33,20 @@ class PostViewSet(viewsets.GenericViewSet):
         data = self.get_serializer(post, many=True).data
         return Response(data, status=status.HTTP_200_OK)
 
-
     @action(detail=False, methods=['GET'], url_path='list')
     def listPosts(self, request): # GET /api/post/list/ | list posts
-        # TODO: implement this function
         try:
             start_num = int(request.data.get('start_num')) # request.query_params.get() ?
             if 'limit_num' in request.data:
                 limit_num = int(request.data.get('limit_num'))
             else:
                 limit_num = 20 # default
-            board_id = int(request.data.get('board')) #TODO: check if board_id valid
+            board_id = int(request.data.get('board'))
+            try:
+                board = Board.objects.get(id=board_id)
+            except ObjectDoesNotExist:
+                return Response({"error": "Board does not exist."}, status=status.HTTP_404_NOT_FOUND)
+
         except TypeError:
             return Response({"error": "start_num, limit_num and board_id must be integers."}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -56,31 +62,37 @@ class PostViewSet(viewsets.GenericViewSet):
     @action(detail=False, methods=['POST'], url_path='write')
     def writePost(self, request): # POST /api/post/write/ | write a post
         user = request.user
+        is_verified = UserProfile.objects.get(user=user).is_verified
+        if not is_verified:
+            return Response({"error": "You do not have a permission to write a post."}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             board_id = int(request.data.get('board'))
         except TypeError:
-            return Response({"error": "board id must be an integer."}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"error": "board_id must be an integer."}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            board = Board.objects.get(id=board_id)
+        except ObjectDoesNotExist:
+            return Response({"error": "Board does not exist."}, status=status.HTTP_404_NOT_FOUND)
+
         serializer = self.get_serializer(data = request.data)
         serializer.is_valid(raise_exception = True)
-        # TODO: title if board.has_title
-        # TODO: check if board_id valid
-
-        # TODO: check permissions
-        #if not user.is_active:
-        #    return Response({"error": "You do not have a permission to write a post."}, status = status.HTTP_403_FORBIDDEN)
-
         serializer.save(user=user, board_id=board_id)
         return Response(serializer.data, status = status.HTTP_201_CREATED)
 
     @transaction.atomic
     @action(detail=True, methods=['PUT'], url_path='like')
     def likePost(self, request, pk=None): # PUT /api/post/{pk}/like/ | like a post
-        # TODO: pk
         user = request.user
+        is_verified = UserProfile.objects.get(user=user).is_verified
+        if not is_verified:
+            return Response({"error": "You do not have a permission to like a post."}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             post = self.get_object()
         except ObjectDoesNotExist:
-            return Response({"error": "Post does not exist."}, status = HTTP_404_NOT_FOUND)
+            return Response({"error": "Post does not exist."}, status = status.HTTP_404_NOT_FOUND)
         try:
             UserLikePost.objects.get(user=user, post=post)
         except ObjectDoesNotExist:
@@ -96,12 +108,18 @@ class PostViewSet(viewsets.GenericViewSet):
 
     @action(detail=True, methods=['DELETE'], url_path='delete')
     def deletePost(self, request, pk=None): # DELETE /api/post/{pk}/delete/ | delete a post
-        # TODO: implement this function
-        user = request.user #TODO: check permission
+        user = request.user
+        is_verified = UserProfile.objects.get(user=user).is_verified
+        if not is_verified:
+            return Response({"error": "You do not have a permission to delete a post."}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             post = self.get_object()
         except ObjectDoesNotExist:
-            return Response({"error": "Post does not exist."}, status = HTTP_404_NOT_FOUND)
+            return Response({"error": "Post does not exist."}, status = status.HTTP_404_NOT_FOUND)
+
+        if user.id != post.user:
+            return Response({"error": "You do not have a permission to delete this post."}, status=status.HTTP_403_FORBIDDEN)
         post.delete()
         return Response(status=status.HTTP_200_OK)
